Changes from PR - CHANGELOG.md file change pending
