+++ sdk/appcenter/src/test/java/com/microsoft/appcenter/utils/InstrumentationRegistryHelperTest.java
@@ -5,12 +5,48 @@
 
 package com.microsoft.appcenter.utils;
 
+import com.microsoft.appcenter.AbstractAppCenterTest;
+import com.microsoft.appcenter.AppCenter;
+import java.lang.reflect.Method;
+import org.junit.Rule;
 import org.junit.Test;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.rule.PowerMockRule;
 
+import static org.junit.Assert.assertFalse;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.when;
+import static org.powermock.api.mockito.PowerMockito.doThrow;
+import static org.powermock.api.mockito.PowerMockito.mockStatic;
+import static org.powermock.api.mockito.PowerMockito.spy;
+import static org.powermock.api.support.membermodification.MemberMatcher.method;
+
+@PrepareForTest({
+        InstrumentationRegistryHelper.class
+})
 public class InstrumentationRegistryHelperTest {
+
+    @Rule
+    public PowerMockRule rule = new PowerMockRule();
 
     @Test
     public void instrumentationRegistryHelperCoverage() {
         new InstrumentationRegistryHelper();
+    }
+
+    @Test
+    public void notRunningInAppCenterTestWhenIllegalStateException() throws Exception {
+        mockStatic(InstrumentationRegistryHelper.class);
+        doThrow(new IllegalStateException()).when(InstrumentationRegistryHelper.class, "getArguments");
+        assertFalse(AppCenter.isRunningInAppCenterTestCloud());
+    }
+
+    @Test
+    public void notRunningInAppCenterTestWhenExceptionInClass() throws Exception {
+        spy(InstrumentationRegistryHelper.class);
+        Method getClass = method(InstrumentationRegistryHelper.class, "getClass", String.class);
+        doThrow(new ClassNotFoundException()).when(InstrumentationRegistryHelper.class, getClass)
+                .withArguments(any(String.class));
+        assertFalse(AppCenter.isRunningInAppCenterTestCloud());
     }
 }
